# Switch from reflect.DeepEqual to bytes.Compare