Results in 4x to 40x speed up checking IsAggregated.

```
benchmark                            old ns/op     new ns/op     delta
BenchmarkIsAggregatedNon_Short-4     167           4.44          -97.34%
BenchmarkIsAggregatedNon_Long-4      165           4.42          -97.32%
BenchmarkIsAggregatedMinimum-4       1642          170           -89.65%
BenchmarkIsAggregatedFull-4          1865          395           -78.82%
```